ajustar as regras de negócio de cart e order (#12)
* ajustar as regras de negocio de cart e order

* ajuste das regras de negocio de cart e order

+++ server/src/use-cases/delete-order/delete-order.ts
@@ -1,4 +1,5 @@
 import { OrderRepository } from '../../repositories/order-repository';
+import { DeleteOrderDTO } from './delete-order-dto';
 
 export class DeleteOrder {
     private orderRepository: OrderRepository;
@@ -8,7 +9,10 @@ export class DeleteOrder {
     }
 
     async execute(id: number): Promise<void> {
-        const order = {
+        if (!(await this.orderRepository.findById(id))) {
+            throw new Error('Order not found');
+        }
+        const order: DeleteOrderDTO = {
             id,
         };
         await this.orderRepository.delete(order.id);
